COD:scheduler refactor: move Google services into package
