+++ tests/services/google/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Google service modules."""
